Include tags for both modbus and generic types
