+++ src/drivers/generic/resources/rest_schema/schema_generic_network.py
@@ -3,9 +3,6 @@ from src.drivers.generic.resources.rest_schema.schema_generic_device import \
 from src.resources.rest_schema.schema_network import *
 
 generic_network_all_attributes = deepcopy(network_all_attributes)
-generic_network_all_attributes['tags'] = {
-    'type': str,
-}
 generic_network_return_attributes = deepcopy(network_return_attributes)
 generic_network_all_attributes['tags'] = {
     'type': str,
