+++ src/models/device/model_device.py
@@ -7,6 +7,7 @@ from src import db
 from src.drivers.enums.drivers import Drivers
 from src.models.model_base import ModelBase
 from src.models.network.model_network import NetworkModel
+from src.utils.model_utils import validate_json
 
 
 class DeviceModel(ModelBase):
@@ -17,6 +18,7 @@ class DeviceModel(ModelBase):
     enable = db.Column(db.Boolean(), nullable=False)
     fault = db.Column(db.Boolean(), nullable=True)
     history_enable = db.Column(db.Boolean(), nullable=False, default=False)
+    tags = db.Column(db.String(320), nullable=True)
     points = db.relationship('PointModel', cascade="all,delete", backref='device', lazy=True)
     driver = db.Column(db.Enum(Drivers), default=Drivers.GENERIC)
 
@@ -31,6 +33,21 @@ class DeviceModel(ModelBase):
 
     def __repr__(self):
         return f"Device(uuid = {self.uuid})"
+
+    @validates('tags')
+    def validate_tags(self, _, value):
+        """
+        Rules for tags:
+        - force all tags to be lower case
+        - if there is a gap add an underscore
+        - no special characters
+        """
+        if value is not None:
+            try:
+                return validate_json(value)
+            except ValueError:
+                raise ValueError('tags needs to be a valid JSON')
+        return value
 
     @validates('name')
     def validate_name(self, _, value):
