+++ src/models/network/model_network.py
@@ -5,6 +5,7 @@ from sqlalchemy.orm import validates
 from src import db
 from src.drivers.enums.drivers import Drivers
 from src.models.model_base import ModelBase
+from src.utils.model_utils import validate_json
 
 
 class NetworkModel(ModelBase):
@@ -14,6 +15,7 @@ class NetworkModel(ModelBase):
     enable = db.Column(db.Boolean(), nullable=False)
     fault = db.Column(db.Boolean(), nullable=True)
     history_enable = db.Column(db.Boolean(), nullable=False, default=False)
+    tags = db.Column(db.String(320), nullable=True)
     devices = db.relationship('DeviceModel', cascade="all,delete", backref='network', lazy=True)
     driver = db.Column(db.Enum(Drivers), default=Drivers.GENERIC)
 
@@ -21,6 +23,21 @@ class NetworkModel(ModelBase):
         'polymorphic_identity': 'network',
         'polymorphic_on': driver
     }
+
+    @validates('tags')
+    def validate_tags(self, _, value):
+        """
+        Rules for tags:
+        - force all tags to be lower case
+        - if there is a gap add an underscore
+        - no special characters
+        """
+        if value is not None:
+            try:
+                return validate_json(value)
+            except ValueError:
+                raise ValueError('tags needs to be a valid JSON')
+        return value
 
     @validates('name')
     def validate_name(self, _, value):
